edit discount info display

+++ app/src/main/java/vn/sharkdms/ui/customer/discount/DiscountInfo.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class DiscountInfo(
     @SerializedName("ruleCode") var ruleCode: String,
     @SerializedName("ruleContent") var ruleContent: String,
-    @SerializedName("minAmount") var minAmount: Int,
-    @SerializedName("maxAmount") var maxAmount: Int,
-    @SerializedName("discountRate") var discountRate: Int
+    @SerializedName("minAmount") var minAmount: Double,
+    @SerializedName("maxAmount") var maxAmount: Double,
+    @SerializedName("discountRate") var discountRate: Double
 )
